Change FFT API to return reals valuesin CCS format
CCS format, is compact format for FFT of real signals.
Because FFT of real signals is symetric, we can express it
in this way:

CCS[0] = real(FFT[0])
CCS[1] = 0
CCS[2] = real(FFT[1])
CCS[3] = imag(FFT[1])
CCS[4] = real(FFT[2])
CCS[5] = imag(FFT[2])
...
CCS[M-1] = real(FFT[N/2])
CCS[M] = 0

Therefore the length of the CCS formated buffer has to be N+2,
if N is the size of the FFT and length of the frame to be transformed.